Build for all archs; Updated build scripts

+++ recipe/build.sh
@@ -4,15 +4,16 @@ IFS=$'\n\t'
 
 mathjax="$PREFIX/lib/mathjax"
 mkdir -p "$mathjax"
+[ $? -ne 0 ] || exit 1
 
-mv config "$mathjax/"
-mv docs "$mathjax/"
-mv extensions "$mathjax/"
-mv fonts "$mathjax/"
-mv jax "$mathjax/"
-mv localization "$mathjax/"
-mv test "$mathjax/"
-mv unpacked "$mathjax/"
+mv config "$mathjax/" || exit 1
+mv docs "$mathjax/" || exit 1
+mv extensions "$mathjax/" || exit 1
+mv fonts "$mathjax/" || exit 1
+mv jax "$mathjax/" || exit 1
+mv localization "$mathjax/" || exit 1
+mv test "$mathjax/" || exit 1
+mv unpacked "$mathjax/" || exit 1
 
 rm *.md ".gitignore" ".travis.yml" "bower.json" "composer.json" "latest.js" "package.json" || exit 1
 cwd="$(pwd)"
